services/visualizer/pyproject.toml
[project]
name = "ocr-visualizer"
version = "0.1.0"
description = "real-time visualization dashboard for ocr pipeline"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "minio>=7.2.18",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "ruff>=0.14.6",
]

[tool.ruff]
line-length = 100
target-version = "py313"
extend-exclude = [
    ".venv",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
# structlog processors
"src/logger.py" = ["ARG001"]
# fastapi lifespan app
"src/main.py" = ["ARG001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
